python-backend/requirements.txt
fastapi==0.135.3
uvicorn==0.42.0
pydantic==2.12.5
# coqui-tts 0.27.4+ 不再内置 PyTorch，需显式安装
torch>=2.2
torchaudio>=2.2.0
coqui-tts==0.27.5
soundfile==0.13.1
pydub==0.25.1
librosa==0.11.0
